requirements.txt
# Core
pyspark>=3.5.0
kafka-python>=2.0.2
delta-spark>=3.0.0
jsonschema>=4.20.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Dashboard
streamlit>=1.29.0
plotly>=5.18.0

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
